moved all questions-related stuff to single file
