+++ src/components/HomePage.js
@@ -1,30 +1,31 @@
 import React from "react";
 import "../styles/HomePage.css";
 import DocumentTitle from "react-document-title";
-import { Link } from 'react-router-dom';
-
-
+import Register from "./Questions";
+import Poll from "./Poll";
 
 export default class HomePage extends React.Component {
+
+    constructor(props) {
+        super(props);
+        this.state = {
+          questions: []
+        };
+      }
+
 	handleClick = () => {
 		sessionStorage.clear();
 		this.props.history.push('/login');
 	}
 
 	render() {
+        console.log(this.state);
 		return (
 			<DocumentTitle title="Home Page">
 				<div className="homePage">
-					<h1>
-						don't leave class confused
-					</h1>
-					<div>
-						<div>
-						</div>
-					</div>
-					<button className="button" onClick={this.handleClick}>Log in</button>
+                <Register />
+                <Poll />
 				</div>
-
 			</DocumentTitle>
 		);
 	}
