+++ src/components/SentimentButton.js
@@ -4,7 +4,7 @@ import Button from 'react-bootstrap/Button';
 const SentimentButton = (props) => {
   return (
     <>
-      <Button variant='primary' name={props.text} onClick={props.onClick} style={{ margin: 8 }}>
+      <Button variant='primary' name={props.text} onClick={props.onClick} style={{ margin: 8 , backgroundColor: "paleturquoise", borderColor:"paleturquoise"}}>
         {props.text}
       </Button>
     </>
