+++ src/index.js
@@ -2,22 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/index.css';
 import App from './App';
+import HomePage from './components/HomePage';
 import * as serviceWorker from './serviceWorker';
 
 import "bootstrap/dist/css/bootstrap.css";
 import "react-notifications/lib/notifications.css";
 
-import { NotificationContainer } from "react-notifications";
-import Register from "./components/Register";
-import SavedList from "./components/SavedList";
-import Poll from "./components/Poll";
-
 ReactDOM.render(
   <React.StrictMode>
-    <NotificationContainer />
-    <SavedList />
-    <Register />
-    <Poll />
+    <HomePage />
   </React.StrictMode>
   ,
   document.getElementById('root')
